signature: Make AuthorIdentity implement {Marshal,Unmarshal}Text
This makes it easier to e.g. put AuthorIdentity values in configuration
and API structs and have the values properly marshaled to/from string.

Since the type has grown we extracted it to a file of its own to keep
the clutter in verifier.go down.
